dev: do not rely on Fusesoc PR
This commit removes a dependency on the `fusesoc core show` command
returning the 'Core file: <name>' string which is currently in a PR.
If the 'show' command doesn't return this string then a default filelist
name of 'filelist.f' is used.  The user also has the option of
overriding the output filename by passing the --output arg.