WIP: provide example of passing backend args
This commit should be modified!

Provides a simple example of passing the backendargs argument to
fusesoc.get_backend.  This would normally be populdated from the fusesoc
command line with any command line args not parsed as fusesoc arguments,
i.e. the other in `args,other = parser.parse_known_args()`.

We can use this to set a tool option that is passed via the EDAM and set
it via command line arguments passed to flist.  We could either check
for tool_<simulator> flags or have another argument to handle this. I
guess there is also the option between passing the simulator name or
just telling the backend to strip the toplevel from the -Gparam=value.