bug/NQ-1354: Correct string parameter passing and parameter scope resolution (#3)
Updates the F-List implementation to be adaptable to different simulator targets. This is necessary because different simulators have different formats for specifying `parameter` and `define` values from the command line. 

To achieve this the F-list command line now takes in an addition `--simulator` arguments which can currently be set to `xcelium` or `verilator`, which is passed to the Fusesoc `backend-args` that are made available to the `Edatool` via `tool_options`. If no argument is specified for this value the verilator format is used, which is somewhat common across a number of open source tools. 

To extend simulator support within F-list extend the `_sim_prefixes` dictionary with aliases for `define` and `param` for the simulator you are targeting. 

---------

Co-authored-by: Shareef Jalloq <shareef.jalloq@nu-quantum.com>