fix: updates for simulator option
This fix updates how we pass backend arguments to the Edatool class.  It
seems that even if you define an argument as having NoneType, if that
tool option has been defined as a string, NoneType gets converted to
"None".  This resulted in different Edatool.get_option() results
depending on whether you ran the flist tool via FuseSoC or via the
command line tool.